Move response queue and satellite API to separate files
